Move more type logic to common module.
