+++ datacapture/build.gradle.kts
@@ -97,6 +97,7 @@ dependencies {
   implementation(Dependencies.Lifecycle.viewModelKtx)
   implementation(Dependencies.material)
   implementation(Dependencies.flexBox)
+  implementation(project(":common"))
 
   testImplementation(Dependencies.AndroidxTest.core)
   testImplementation(Dependencies.junit)
