+++ datacapture/src/main/java/com/google/android/fhir/datacapture/validation/MaxValueConstraintValidator.kt
@@ -17,6 +17,7 @@
 package com.google.android.fhir.datacapture.validation
 
 import android.content.Context
+import com.google.android.fhir.compareTo
 import com.google.android.fhir.datacapture.R
 import org.hl7.fhir.r4.model.Extension
 import org.hl7.fhir.r4.model.QuestionnaireResponse
