+++ datacapture/src/main/java/com/google/android/fhir/datacapture/validation/MinValueConstraintValidator.kt
@@ -17,6 +17,7 @@
 package com.google.android.fhir.datacapture.validation
 
 import android.content.Context
+import com.google.android.fhir.compareTo
 import com.google.android.fhir.datacapture.R
 import org.hl7.fhir.r4.model.Extension
 import org.hl7.fhir.r4.model.QuestionnaireResponse
@@ -29,6 +30,7 @@ internal object MinValueConstraintValidator :
       extension: Extension,
       answer: QuestionnaireResponse.QuestionnaireResponseItemAnswerComponent ->
       answer.value < extension.value
+      true
     },
     { extension: Extension, context: Context ->
       context.getString(R.string.min_value_validation_error_msg, extension.value.primitiveValue())
